+++ R/test-files.R
@@ -67,16 +67,10 @@ test_dir <- function(
     cli::cli_abort("No test files found.")
   }
 
-  want_parallel <- find_parallel(path, load_package, package)
+  want_parallel <- find_parallel(path, load_package, package) &&
+    length(test_paths) > 1
 
-  if (is.null(reporter)) {
-    if (want_parallel) {
-      reporter <- default_parallel_reporter()
-    } else {
-      reporter <- default_reporter()
-    }
-  }
-  reporter <- find_reporter(reporter)
+  reporter <- find_reporter(reporter %||% default_reporter(want_parallel))
   parallel <- want_parallel && reporter$capabilities$parallel_support
 
   test_files(