+++ tests/testthat/test-parallel/teardown/tests/testthat/test-empty.R
@@ -1,0 +1,1 @@
+# Empty test (used to ensure parallel execution is performed)